rename helper function to filter Quiz Stats
