+++ controllers/quizes/findAllQuizzes.js
@@ -1,9 +1,9 @@
 const db = require("../../models");
-const filterQuestionObjs = require("../../helpers/filterQuestionObjs/filterQuestionObjs");
+const filterQuizStats = require("../../helpers/filterQuizStats/filterQuizStats");
 
 module.exports = function(req,res){
   db.Quiz.findAll({}).then(result => {
-    res.json(filterQuestionObjs(result));
+    res.json(filterQuizStats(result));
   }).catch(function(err){
     res.json(err);
   });
